Updated module augmentation for cloud-services package.
