+++ packages/ckeditor5-build-balloon/src/ckeditor.ts
@@ -30,7 +30,7 @@ import PasteFromOffice from '@ckeditor/ckeditor5-paste-from-office/src/pastefrom
 import PictureEditing from '@ckeditor/ckeditor5-image/src/pictureediting';
 import { Table, TableToolbar } from '@ckeditor/ckeditor5-table';
 import TextTransformation from '@ckeditor/ckeditor5-typing/src/texttransformation';
-import CloudServices from '@ckeditor/ckeditor5-cloud-services/src/cloudservices';
+import { CloudServices } from '@ckeditor/ckeditor5-cloud-services';
 
 export default class BalloonEditor extends BalloonEditorBase {}
