+++ packages/ckeditor5-build-decoupled-document/src/ckeditor.ts
@@ -40,7 +40,7 @@ import PasteFromOffice from '@ckeditor/ckeditor5-paste-from-office/src/pastefrom
 import PictureEditing from '@ckeditor/ckeditor5-image/src/pictureediting';
 import { Table, TableToolbar } from '@ckeditor/ckeditor5-table';
 import TextTransformation from '@ckeditor/ckeditor5-typing/src/texttransformation';
-import CloudServices from '@ckeditor/ckeditor5-cloud-services/src/cloudservices';
+import { CloudServices } from '@ckeditor/ckeditor5-cloud-services';
 
 export default class DecoupledEditor extends DecoupledEditorBase {}
